+++ Example/Tests/SpireTests.m
@@ -90,6 +90,7 @@
     if (!saleResponse)
     {
         XCTFail(@"Sale did not succeed. Make sure your terminal is paired in your iOS device settings and that the terminal is in stand-by mode (ie. by switching off and then on and waiting until the screen lights off).");
+        //NOTE:  if your merchant settings have cash mgmt enabled in backend, you will need to run cash tests first - otherwise you will receive a "not authorized" kind of error
     }
     else
     {